+++ packages/codemod/src/spinner/__testfixtures__/transform.output.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+import { Spinner } from '@alfalab/core-components/spinner';
+
+export const Component = () => {
+    const someProps = { size: 48 };
+
+    return (
+        <React.Fragment>
+            <Spinner preset={16} />
+            <Spinner preset={24} />
+            <Spinner preset={48} />
+            <Spinner preset={16} />
+            <Spinner preset={24} />
+            <Spinner preset={48} />
+            <Spinner visible={true} />
+            <Spinner {...someProps} />
+        </React.Fragment>
+    );
+};