+++ packages/date-time-input/src/Component.responsive.tsx
@@ -2,8 +2,7 @@ import React, { forwardRef } from 'react';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { InputProps } from '@alfalab/core-components-input';
-import { getComponentBreakpoint } from '@alfalab/core-components-shared';
-import { useMedia } from '@alfalab/hooks';
+import { useIsDesktop } from '@alfalab/core-components-mq';
 
 import { DateTimeInputProps } from './components/date-time-input/Component';
 import { DateTimeInputDesktop } from './desktop';
@@ -15,21 +14,20 @@ export type DateTimeInputResponsiveProps = Omit<DateTimeInputProps, 'view' | 'In
      * @default 1024
      */
     breakpoint?: number;
+
+    /**
+     * Версия, которая будет использоваться при серверном рендеринге
+     */
+    client?: 'desktop' | 'mobile';
 };
 
 export type DateTimeInputMedia = 'desktop' | 'mobile';
 
 export const DateTimeInputResponsive = forwardRef<HTMLInputElement, DateTimeInputResponsiveProps>(
-    ({ breakpoint = getComponentBreakpoint(), ...restProps }, ref) => {
-        const [view] = useMedia<DateTimeInputMedia>(
-            [
-                ['mobile', `(max-width: ${breakpoint - 1}px)`],
-                ['desktop', `(min-width: ${breakpoint}px)`],
-            ],
-            'desktop',
-        );
+    ({ breakpoint, client, ...restProps }, ref) => {
+        const isDesktop = useIsDesktop(breakpoint, client === 'desktop');
 
-        return view === 'desktop' ? (
+        return isDesktop ? (
             <DateTimeInputDesktop {...restProps} ref={ref} />
         ) : (
             <DateTimeInputMobile {...restProps} ref={ref} />