+++ packages/file-upload-item/src/components/actions-control/components/restore-button/index.ts
@@ -1,0 +1,1 @@
+export { RestoreButton } from './restore-button';