+++ packages/file-upload-item/src/const/progress-bar.ts
@@ -1,0 +1,2 @@
+export const MAX_PROGRESS_BAR_VALUE = 100;
+export const RADIUS = 360;