+++ packages/international-phone-input/src/types.ts
@@ -127,7 +127,13 @@ export type InternationalPhoneInputProps = InternationalPhoneInputDesktopProps &
         breakpoint?: number;
 
         /**
+         * Версия, которая будет использоваться при серверном рендеринге
+         */
+        client?: 'desktop' | 'mobile';
+
+        /**
          * Значение по-умолчанию для хука useMatchMedia
+         * @deprecated Используйте client
          */
         defaultMatchMediaValue?: boolean | (() => boolean);
     };