+++ packages/slider-input/src/types/propTypes.ts
@@ -1,0 +1,11 @@
+import { ChangeEvent } from 'react';
+
+export type OnChangeType = (
+    event: ChangeEvent<HTMLInputElement> | null,
+    payload: { value: number | '' },
+) => void;
+
+export type OnInputChangeType = (
+    event: ChangeEvent<HTMLInputElement> | null,
+    payload: { value: number | '' },
+) => void;