+++ packages/tooltip/src/Component.responsive.tsx
@@ -1,21 +1,21 @@
 import React, { FC } from 'react';
 
-import { useMatchMedia } from '@alfalab/core-components-mq';
-import { getComponentBreakpoint } from '@alfalab/core-components-shared';
+import { useIsDesktop } from '@alfalab/core-components-mq';
 
 import { TooltipDesktop } from './desktop';
 import { TooltipMobile } from './mobile';
 import { TooltipResponsiveProps } from './types';
 
 export const TooltipResponsive: FC<TooltipResponsiveProps> = ({
-    defaultMatchMediaValue,
     children,
     actionButtonTitle,
     bottomSheetProps,
-    breakpoint = getComponentBreakpoint(),
+    breakpoint,
+    client,
+    defaultMatchMediaValue = client === undefined ? undefined : client === 'desktop',
     ...restProps
 }) => {
-    const [isDesktop] = useMatchMedia(`(min-width: ${breakpoint}px)`, defaultMatchMediaValue);
+    const isDesktop = useIsDesktop(breakpoint, defaultMatchMediaValue);
 
     return isDesktop ? (
         <TooltipDesktop {...restProps}>{children}</TooltipDesktop>