+++ packages/custom-picker-button/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @alfalab/core-components-custom-picker-button
 
+## 2.6.4
+
+### Patch Changes
+
+<sup><time>18.11.2024</time></sup>
+
+### [#1426](https://github.com/core-ds/core-components/pull/1426)
+
+-   Добавлен пакет @alfalab/core-config для глобальных настроек библиотеки. В него включены параметры: breakpoint для переключения между десктопной и мобильной версиями и client для выбора версии по умолчанию при серверном рендеринге
+
+-   Обновлены зависимости
+    -   shared@0.14.0
+    -   custom-button@3.10.1
+    -   picker-button@11.12.4
+    -   select@17.17.2
+
 ## 2.6.3
 
 ### Patch Changes