+++ packages/markdown/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @alfalab/core-components-markdown
 
+## 1.6.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.14.0
+
 ## 1.6.0
 
 ### Minor Changes