+++ packages/masked-input/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 6.3.26
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@15.3.1
+
 ## 6.3.25
 
 ### Patch Changes