+++ packages/password-input/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 5.6.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@15.3.1
+    -   icon-button@6.11.1
+
 ## 5.6.1
 
 ### Patch Changes