+++ packages/tabs/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 8.12.4
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.14.0
+    -   picker-button@11.12.4
+    -   tag@8.6.1
+    -   icon-button@6.11.1
+
 ## 8.12.3
 
 ### Patch Changes