+++ packages/with-suffix/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 4.2.7
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@15.3.1
+    -   portal@3.3.4
+
 ## 4.2.6
 
 ### Patch Changes